rag_project/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
numpy
Pillow
pypdf
pymupdf
pytesseract
pdf2image
faiss-cpu
huggingface-hub
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
